requirements.txt
aiogram==2.9.0
aiohttp
aiosqlite
asyncpg
Pyrogram
PySocks
pytest
pytz
tortoise-orm
python-dotenv
